trunk: offsets within the domain promoted to integer(kind=8)

+++ problems/sedov/initproblem.F90
@@ -193,7 +193,7 @@ contains
 
       character(len=*), intent(in)        :: var   !< quantity to be plotted
       character(len=*), intent(in)        :: ij    !< plane of plot
-      integer, intent(in)                 :: xn    !< no. of cell at which we are slicing the local block
+      integer(kind=8), intent(in)         :: xn    !< no. of cell at which we are slicing the local block
       integer, intent(inout)              :: ierrh !< error handling
       real, dimension(:,:), intent(inout) :: tab   !< array  containing given quantity
